+++ changelog.md
@@ -8,6 +8,7 @@
 - Reduced the published size of `package.json` by moving dev tool configs to separate files.
 - Remove `.html` from [usejsdoc.org](http://usejsdoc.org) links.
 - Test `outlineMembers` with missing members.
+- Test `membersToMdAst` with a parameter default value.
 
 ## 2.1.0